feat: Bearer 인증 기능 추가 및 테스트, 깨지는 테스트 수정
